requirements.txt
alabaster==0.7.13
Babel==2.11.0
beautifulsoup4==4.11.1
certifi==2022.12.7
charset-normalizer==2.0.0
click==8.0.4
contourpy==1.0.6
cycler==0.11.0
cython==0.29.33
docutils==0.17.1
fonttools==4.38.0
furo==2022.9.29
greenlet==2.0.1
idna==3.4
imagecodecs==2021.4.28
imageio==2.23.0
imagesize==1.4.1
imohash==1.0.4
importlib-metadata==6.0.0
Jinja2==3.1.2
kiwisolver==1.4.4
markdown-it-py==2.1.0
MarkupSafe==2.1.2
matplotlib==3.6.2
mdit-py-plugins==0.3.3
mdurl==0.1.2
mmh3==3.0.0
networkx==2.8.8
numpy==1.24.1
numpydoc==1.5.0
opencv-python==4.7.0.68
packaging==22.0
Pillow==9.3.0
Pygments==2.14.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
PyWavelets==1.4.1
PyYAML==6.0
rawpy==0.17.3
requests==2.26.0
scikit-image==0.19.3
scipy==1.9.3
six==1.16.0
snowballstemmer==2.2.0
soupsieve==2.3.2.post1
Sphinx==4.5.0
sphinx-basic-ng==1.0.0b1
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
SQLAlchemy==1.4.44
SQLAlchemy-Utils==0.38.3
tifffile==2022.10.10
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.14
varint==1.0.2
watchdog==2.2.1
zipp==3.14.0
